Add fixed footer with copyright information to errorlogin.html, index.html, and login.html
This commit adds a fixed footer to the bottom of the errorlogin.html, index.html, and login.html pages. The footer contains a copyright notice and the name of the application (FRÖCCS ONLINE by LevyProductions) along with its version number (v1.1). The footer has a gray background and white text, and is centered horizontally.